pyproject.toml
[tool.poetry]
name = "ts_scaler"
version = "0.0.1"
description = ""
authors = [
    "Joseph Wang <egpivo@gmail.com>",
    "Hsin-Cheng Huang <hchuang@stat.sinica.edu.tw>",
    "ShengLi Tzeng <slt.cmu@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
ipykernel = "6.13.0"
ipython = "8.3.0"
boto3 = "^1.34.153"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.971"
pydocstyle = { extras = ["toml"], version = "^6.1.1" }
pytest = "^7.1.3"
pytest-cov = "^4.1.0"
poethepoet = "^0.15.0"
coverage = { extras = ["toml"], version = "^6.4.4" }
pre-commit = "^2.20.0"
python-dotenv = "^1.0.0"
poetry-bumpversion = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = "cleo.*"
ignore_missing_imports = true
